# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atlas of AI — Editorial Desk API",
    "version": "0.1.0",
    "description": "CRUD over the AI-models JSON store. Every mutation also regenerates the public GeoJSON file consumed by the static viewer."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "schemas": {
      "Model": {
        "type": "object",
        "required": ["name", "organization", "lat", "lng"],
        "properties": {
          "id":           { "type": "string", "description": "Auto-derived from name if omitted" },
          "name":         { "type": "string" },
          "organization": { "type": "string" },
          "country":      { "type": "string" },
          "city":         { "type": "string" },
          "lat":          { "type": "number", "minimum": -90,  "maximum": 90 },
          "lng":          { "type": "number", "minimum": -180, "maximum": 180 },
          "type":         { "type": "string", "enum": ["proprietary", "open-weight"] },
          "year":         { "type": "integer" },
          "parameters":   { "type": "string" },
          "url":          { "type": "string" },
          "notes":        { "type": "string" },
          "modality":     { "type": "array", "items": { "type": "string", "enum": ["text", "image", "audio", "video", "code", "3d"] } }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "edition":  { "type": "string" },
          "updated":  { "type": "string", "format": "date" },
          "compiler": { "type": "string" }
        }
      },
      "Edition": {
        "type": "object",
        "properties": {
          "id":       { "type": "string", "description": "Unique edition id, e.g. 2026-04-19_vol-i-mmxxvi-abc123" },
          "label":    { "type": "string", "description": "Human-readable edition title" },
          "note":     { "type": "string", "description": "Editorial note" },
          "date":     { "type": "string", "format": "date" },
          "features": { "type": "integer", "description": "Number of models snapshotted" },
          "file":     { "type": "string", "description": "Path relative to the data root, e.g. editions/2026-04-19_….geojson" }
        }
      },
      "Tool": {
        "type": "object",
        "required": ["name", "organization", "lat", "lng"],
        "properties": {
          "id":           { "type": "string", "description": "Auto-derived from name if omitted" },
          "name":         { "type": "string" },
          "organization": { "type": "string" },
          "country":      { "type": "string" },
          "city":         { "type": "string" },
          "lat":          { "type": "number", "minimum": -90,  "maximum": 90 },
          "lng":          { "type": "number", "minimum": -180, "maximum": 180 },
          "category":     { "type": "string", "enum": ["assistant", "search", "ide", "codegen", "devtool", "text-to-image", "text-to-speech", "text-to-video", "music","local LLM tool"], "default": "devtool" },
          "year":         { "type": "integer" },
          "url":          { "type": "string" },
          "notes":        { "type": "string" },
          "builtOn":      { "type": "array", "items": { "type": "string" }, "description": "Model ids this tool is built on" }
        }
      },
      "GlossaryTerm": {
        "type": "object",
        "required": ["term", "definition"],
        "properties": {
          "id":         { "type": "string", "description": "Auto-derived from term if omitted" },
          "term":       { "type": "string" },
          "definition": { "type": "string" },
          "notes":      { "type": "string" },
          "links":        { "type": "array", "items": { "type": "object", "properties": { "label": { "type": "string" }, "url": { "type": "string" } } } },
          "relatedTerms": { "type": "array", "items": { "type": "string" }, "description": "IDs of related glossary terms" },
          "updatedAt":    { "type": "string", "format": "date" }
        }
      },
      "GlossaryEdition": {
        "type": "object",
        "properties": {
          "id":     { "type": "string" },
          "label":  { "type": "string" },
          "note":   { "type": "string" },
          "date":   { "type": "string", "format": "date" },
          "terms":  { "type": "integer", "description": "Number of terms snapshotted" },
          "file":   { "type": "string", "description": "Path relative to the data root, e.g. glossary-editions/2026-04-28_….json" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error":   { "type": "string" },
          "details": { "type": "array", "items": { "type": "string" } }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness probe",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/meta": {
      "get": { "summary": "Get atlas meta", "responses": { "200": { "description": "Meta", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Meta" } } } } } },
      "put": {
        "summary": "Update atlas meta",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Meta" } } } },
        "responses": { "200": { "description": "Updated meta" } }
      }
    },
    "/models": {
      "get": {
        "summary": "List all models",
        "responses": {
          "200": {
            "description": "Array of models",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "models": { "type": "array", "items": { "$ref": "#/components/schemas/Model" } } } } } }
          }
        }
      },
      "post": {
        "summary": "Create a model",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } } },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "ID conflict" }
        }
      }
    },
    "/models/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get":    { "summary": "Get one model", "responses": { "200": { "description": "Model" }, "404": { "description": "Not found" } } },
      "put":    {
        "summary": "Update a model",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": { "summary": "Delete a model", "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
    "/export": {
      "post": {
        "summary": "Force regeneration of the public GeoJSON file",
        "responses": { "200": { "description": "Wrote file" } }
      }
    },
    "/editions": {
      "get": {
        "summary": "List all saved editions",
        "responses": {
          "200": {
            "description": "Array of edition metadata",
            "content": { "application/json": { "schema": {
              "type": "object",
              "properties": { "editions": { "type": "array", "items": { "$ref": "#/components/schemas/Edition" } } }
            }}}
          }
        }
      },
      "post": {
        "summary": "Save current catalogue as a named edition",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["label"],
            "properties": {
              "label": { "type": "string", "description": "Edition title, e.g. Vol. II — Autumn 2026" },
              "note":  { "type": "string", "description": "Optional editorial note" }
            }
          }}}
        },
        "responses": {
          "201": { "description": "Created edition metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Edition" } } } },
          "400": { "description": "Label missing" }
        }
      }
    },
    "/editions/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "summary": "Get GeoJSON FeatureCollection for a specific edition",
        "responses": {
          "200": { "description": "GeoJSON FeatureCollection" },
          "404": { "description": "Edition not found" }
        }
      }
    },
    "/tools": {
      "get": {
        "summary": "List all tools",
        "responses": {
          "200": {
            "description": "Array of tools",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "tools": { "type": "array", "items": { "$ref": "#/components/schemas/Tool" } } } } } }
          }
        }
      },
      "post": {
        "summary": "Create a tool",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tool" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tool" } } } },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "ID conflict" }
        }
      }
    },
    "/tools/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get":    { "summary": "Get one tool", "responses": { "200": { "description": "Tool" }, "404": { "description": "Not found" } } },
      "put":    {
        "summary": "Update a tool",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tool" } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": { "summary": "Delete a tool", "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
    "/glossary-editions": {
      "get": {
        "summary": "List all glossary editions",
        "responses": { "200": { "description": "Array of edition metadata", "content": { "application/json": { "schema": { "type": "object", "properties": { "editions": { "type": "array", "items": { "$ref": "#/components/schemas/GlossaryEdition" } } } } } } } }
      },
      "post": {
        "summary": "Save current glossary as a named edition",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["label"], "properties": { "label": { "type": "string" }, "note": { "type": "string" } } } } } },
        "responses": {
          "201": { "description": "Created edition metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GlossaryEdition" } } } },
          "400": { "description": "Label missing" }
        }
      }
    },
    "/glossary-editions/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get":    { "summary": "Get snapshot for a glossary edition", "responses": { "200": { "description": "Glossary snapshot" }, "404": { "description": "Not found" } } },
      "put":    { "summary": "Update glossary edition label/note", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GlossaryEdition" } } } }, "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } } },
      "delete": { "summary": "Delete a glossary edition", "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
    "/glossary": {
      "get": {
        "summary": "List all glossary terms (alphabetical)",
        "responses": {
          "200": {
            "description": "Array of terms",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "terms": { "type": "array", "items": { "$ref": "#/components/schemas/GlossaryTerm" } } } } } }
          }
        }
      },
      "post": {
        "summary": "Create a glossary term",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GlossaryTerm" } } } },
        "responses": {
          "201": { "description": "Created term" },
          "400": { "description": "Validation error" },
          "409": { "description": "ID conflict" }
        }
      }
    },
    "/glossary/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get":    { "summary": "Get one glossary term", "responses": { "200": { "description": "Term" }, "404": { "description": "Not found" } } },
      "put":    {
        "summary": "Update a glossary term",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GlossaryTerm" } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": { "summary": "Delete a glossary term", "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    }
  }
}